pyproject.toml
[tool.poetry]
name = "telecog-provider-interface"
version = "0.1.0"
description = ""
authors = ["kan kawabata <kkawabat@asu.edu>"]
readme = "README.md"
packages = [{include = "*", from="src"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.7"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
twilio = "^9.2.3"
python-dotenv = "^1.0.1"
short-url = "^1.2.2"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
